Create UCSB-CHG_CHIRTS_DAILY_jsonnet
CHIRTS temperature products are designed to support analysis of temperature extremes and variability, especially in regions with a low density of station observations, which include large areas of the Global South. 

This jsonnet is for the CHIRTS-daily dataset, which provides high resolution (0.05 x 0.05 degree resolution) gridded daily maximum and minimum temperatures and several related variables. The data extent is quasi-global (60°S – 70°N) and currently covers the period 1983-2016. CHIRTS-daily is a unique hybrid of observations from stations and satellites, and estimates from reanalysis data (ECMWF ERA5)

The Climate Hazards Center at the University of California, Santa Barbara (https://chc.ucsb.edu/) is the data provider of CHIRTS and also the CHIRPS product (currently in the earth engine catalog). CHC is currently listed in the earth engine catalog as CHG. 

CHC has an active role in operational agroclimatology monitoring and forecasting to support early warning about food security impacts, conducts research on climate extremes and crop and health impacts. CHC produces unique datasets that are valuable to the scientific and climate impact analysis community.